+++ src/alf/atoms.ts
@@ -300,6 +300,9 @@ export const atoms = {
   /*
    * Padding
    */
+  p_0: {
+    padding: 0,
+  },
   p_2xs: {
     padding: tokens.space._2xs,
   },
@@ -329,6 +332,10 @@ export const atoms = {
   },
   p_5xl: {
     padding: tokens.space._5xl,
+  },
+  px_0: {
+    paddingLeft: 0,
+    paddingRight: 0,
   },
   px_2xs: {
     paddingLeft: tokens.space._2xs,
@@ -370,6 +377,10 @@ export const atoms = {
     paddingLeft: tokens.space._5xl,
     paddingRight: tokens.space._5xl,
   },
+  py_0: {
+    paddingTop: 0,
+    paddingBottom: 0,
+  },
   py_2xs: {
     paddingTop: tokens.space._2xs,
     paddingBottom: tokens.space._2xs,
@@ -410,6 +421,9 @@ export const atoms = {
     paddingTop: tokens.space._5xl,
     paddingBottom: tokens.space._5xl,
   },
+  pt_0: {
+    paddingTop: 0,
+  },
   pt_2xs: {
     paddingTop: tokens.space._2xs,
   },
@@ -439,6 +453,9 @@ export const atoms = {
   },
   pt_5xl: {
     paddingTop: tokens.space._5xl,
+  },
+  pb_0: {
+    paddingBottom: 0,
   },
   pb_2xs: {
     paddingBottom: tokens.space._2xs,
@@ -470,6 +487,9 @@ export const atoms = {
   pb_5xl: {
     paddingBottom: tokens.space._5xl,
   },
+  pl_0: {
+    paddingLeft: 0,
+  },
   pl_2xs: {
     paddingLeft: tokens.space._2xs,
   },
@@ -499,6 +519,9 @@ export const atoms = {
   },
   pl_5xl: {
     paddingLeft: tokens.space._5xl,
+  },
+  pr_0: {
+    paddingRight: 0,
   },
   pr_2xs: {
     paddingRight: tokens.space._2xs,
@@ -534,6 +557,9 @@ export const atoms = {
   /*
    * Margin
    */
+  m_0: {
+    margin: 0,
+  },
   m_2xs: {
     margin: tokens.space._2xs,
   },
@@ -563,6 +589,13 @@ export const atoms = {
   },
   m_5xl: {
     margin: tokens.space._5xl,
+  },
+  m_auto: {
+    margin: 'auto',
+  },
+  mx_0: {
+    marginLeft: 0,
+    marginRight: 0,
   },
   mx_2xs: {
     marginLeft: tokens.space._2xs,
@@ -604,6 +637,14 @@ export const atoms = {
     marginLeft: tokens.space._5xl,
     marginRight: tokens.space._5xl,
   },
+  mx_auto: {
+    marginLeft: 'auto',
+    marginRight: 'auto',
+  },
+  my_0: {
+    marginTop: 0,
+    marginBottom: 0,
+  },
   my_2xs: {
     marginTop: tokens.space._2xs,
     marginBottom: tokens.space._2xs,
@@ -644,6 +685,13 @@ export const atoms = {
     marginTop: tokens.space._5xl,
     marginBottom: tokens.space._5xl,
   },
+  my_auto: {
+    marginTop: 'auto',
+    marginBottom: 'auto',
+  },
+  mt_0: {
+    marginTop: 0,
+  },
   mt_2xs: {
     marginTop: tokens.space._2xs,
   },
@@ -673,6 +721,12 @@ export const atoms = {
   },
   mt_5xl: {
     marginTop: tokens.space._5xl,
+  },
+  mt_auto: {
+    marginTop: 'auto',
+  },
+  mb_0: {
+    marginBottom: 0,
   },
   mb_2xs: {
     marginBottom: tokens.space._2xs,
@@ -704,6 +758,12 @@ export const atoms = {
   mb_5xl: {
     marginBottom: tokens.space._5xl,
   },
+  mb_auto: {
+    marginBottom: 'auto',
+  },
+  ml_0: {
+    marginLeft: 0,
+  },
   ml_2xs: {
     marginLeft: tokens.space._2xs,
   },
@@ -734,6 +794,12 @@ export const atoms = {
   ml_5xl: {
     marginLeft: tokens.space._5xl,
   },
+  ml_auto: {
+    marginLeft: 'auto',
+  },
+  mr_0: {
+    marginRight: 0,
+  },
   mr_2xs: {
     marginRight: tokens.space._2xs,
   },
@@ -763,5 +829,8 @@ export const atoms = {
   },
   mr_5xl: {
     marginRight: tokens.space._5xl,
+  },
+  mr_auto: {
+    marginRight: 'auto',
   },
 } as const
