+++ src/screens/Login/LoginForm.tsx
@@ -2,36 +2,29 @@ import React, {useState, useRef} from 'react'
 import {
   ActivityIndicator,
   Keyboard,
-  ScrollView,
   TextInput,
   TouchableOpacity,
   View,
 } from 'react-native'
 import {ComAtprotoServerDescribeServer} from '@atproto/api'
 import {Trans, msg} from '@lingui/macro'
+import {useLingui} from '@lingui/react'
 
 import {useAnalytics} from 'lib/analytics/analytics'
-import {s} from 'lib/styles'
 import {createFullHandle} from 'lib/strings/handles'
-import {toNiceDomain} from 'lib/strings/url-helpers'
 import {isNetworkError} from 'lib/strings/errors'
 import {useSessionApi} from '#/state/session'
 import {cleanError} from 'lib/strings/errors'
 import {logger} from '#/logger'
-import {styles} from '../../view/com/auth/login/styles'
-import {useLingui} from '@lingui/react'
-import {useDialogControl} from '#/components/Dialog'
-import {ServerInputDialog} from '../../view/com/auth/server-input'
 import {Button, ButtonText} from '#/components/Button'
-import {isAndroid} from '#/platform/detection'
-import {atoms as a, useBreakpoints, useTheme} from '#/alf'
+import {atoms as a, useTheme} from '#/alf'
 import {Text} from '#/components/Typography'
 import * as TextField from '#/components/forms/TextField'
 import {At_Stroke2_Corner0_Rounded as At} from '#/components/icons/At'
 import {Lock_Stroke2_Corner0_Rounded as Lock} from '#/components/icons/Lock'
-import {Globe_Stroke2_Corner0_Rounded as Globe} from '#/components/icons/Globe'
-import {Pencil_Stroke2_Corner0_Rounded as Pencil} from '#/components/icons/Pencil'
-import {Warning_Stroke2_Corner0_Rounded as Warning} from '#/components/icons/Warning'
+import {HostingProvider} from '#/components/forms/HostingProvider'
+import {FormContainer} from './FormContainer'
+import {FormError} from './FormError'
 
 type ServiceDescription = ComAtprotoServerDescribeServer.OutputSchema
 
@@ -64,14 +57,11 @@ export const LoginForm = ({
   const passwordInputRef = useRef<TextInput>(null)
   const {_} = useLingui()
   const {login} = useSessionApi()
-  const serverInputControl = useDialogControl()
-  const {gtMobile} = useBreakpoints()
 
-  const onPressSelectService = () => {
-    serverInputControl.open()
+  const onPressSelectService = React.useCallback(() => {
     Keyboard.dismiss()
     track('Signin:PressedSelectService')
-  }
+  }, [track])
 
   const onPressNext = async () => {
     Keyboard.dismiss()
@@ -131,171 +121,138 @@ export const LoginForm = ({
 
   const isReady = !!serviceDescription && !!identifier && !!password
   return (
-    <ScrollView testID="loginForm" style={a.h_full}>
-      <View style={[a.gap_lg, !gtMobile && a.px_lg, a.flex_1]}>
-        <ServerInputDialog
-          control={serverInputControl}
-          onSelect={setServiceUrl}
+    <FormContainer testID="loginForm" title={<Trans>Sign in</Trans>}>
+      <View>
+        <TextField.Label>
+          <Trans>Hosting provider</Trans>
+        </TextField.Label>
+        <HostingProvider
+          serviceUrl={serviceUrl}
+          onSelectServiceUrl={setServiceUrl}
+          onOpenDialog={onPressSelectService}
         />
-
-        <View>
-          <TextField.Label>
-            <Trans>Hosting provider</Trans>
-          </TextField.Label>
+      </View>
+      <View>
+        <TextField.Label>
+          <Trans>Account</Trans>
+        </TextField.Label>
+        <TextField.Root>
+          <TextField.Icon icon={At} />
+          <TextField.Input
+            testID="loginUsernameInput"
+            label={_(msg`Username or email address`)}
+            autoCapitalize="none"
+            autoFocus
+            autoCorrect={false}
+            autoComplete="username"
+            returnKeyType="next"
+            textContentType="username"
+            onSubmitEditing={() => {
+              passwordInputRef.current?.focus()
+            }}
+            blurOnSubmit={false} // prevents flickering due to onSubmitEditing going to next field
+            value={identifier}
+            onChangeText={str =>
+              setIdentifier((str || '').toLowerCase().trim())
+            }
+            editable={!isProcessing}
+            accessibilityHint={_(
+              msg`Input the username or email address you used at signup`,
+            )}
+          />
+        </TextField.Root>
+      </View>
+      <View>
+        <TextField.Root>
+          <TextField.Icon icon={Lock} />
+          <TextField.Input
+            testID="loginPasswordInput"
+            inputRef={passwordInputRef}
+            label={_(msg`Password`)}
+            autoCapitalize="none"
+            autoCorrect={false}
+            autoComplete="password"
+            returnKeyType="done"
+            enablesReturnKeyAutomatically={true}
+            secureTextEntry={true}
+            textContentType="password"
+            clearButtonMode="while-editing"
+            value={password}
+            onChangeText={setPassword}
+            onSubmitEditing={onPressNext}
+            blurOnSubmit={false} // HACK: https://github.com/facebook/react-native/issues/21911#issuecomment-558343069 Keyboard blur behavior is now handled in onSubmitEditing
+            editable={!isProcessing}
+            accessibilityHint={
+              identifier === ''
+                ? _(msg`Input your password`)
+                : _(msg`Input the password tied to ${identifier}`)
+            }
+          />
           <TouchableOpacity
+            testID="forgotPasswordButton"
+            onPress={onPressForgotPassword}
             accessibilityRole="button"
+            accessibilityLabel={_(msg`Forgot password`)}
+            accessibilityHint={_(msg`Opens password reset form`)}
             style={[
-              a.w_full,
-              a.flex_row,
-              a.align_center,
               a.rounded_sm,
-              a.px_md,
-              a.gap_xs,
-              {paddingVertical: isAndroid ? 14 : 9},
-              t.atoms.bg_contrast_25,
-            ]}
-            onPress={onPressSelectService}>
-            <TextField.Icon icon={Globe} />
-            <Text style={[a.text_md]}>{toNiceDomain(serviceUrl)}</Text>
-            <View
-              style={[
-                a.rounded_sm,
-                t.atoms.bg_contrast_100,
-                {marginLeft: 'auto', left: 6, padding: 6},
-              ]}>
-              <Pencil
-                style={{color: t.palette.contrast_500}}
-                height={18}
-                width={18}
-              />
-            </View>
+              t.atoms.bg_contrast_100,
+              {marginLeft: 'auto', left: 6, padding: 6},
+              a.z_10,
+            ]}>
+            <ButtonText style={t.atoms.text_contrast_medium}>
+              <Trans>Forgot?</Trans>
+            </ButtonText>
           </TouchableOpacity>
-        </View>
-        <View>
-          <TextField.Label>
-            <Trans>Account</Trans>
-          </TextField.Label>
-          <TextField.Root>
-            <TextField.Icon icon={At} />
-            <TextField.Input
-              testID="loginUsernameInput"
-              label={_(msg`Username or email address`)}
-              autoCapitalize="none"
-              autoFocus
-              autoCorrect={false}
-              autoComplete="username"
-              returnKeyType="next"
-              textContentType="username"
-              onSubmitEditing={() => {
-                passwordInputRef.current?.focus()
-              }}
-              blurOnSubmit={false} // prevents flickering due to onSubmitEditing going to next field
-              value={identifier}
-              onChangeText={str =>
-                setIdentifier((str || '').toLowerCase().trim())
-              }
-              editable={!isProcessing}
-              accessibilityHint={_(
-                msg`Input the username or email address you used at signup`,
-              )}
-            />
-          </TextField.Root>
-        </View>
-        <View>
-          <TextField.Root>
-            <TextField.Icon icon={Lock} />
-            <TextField.Input
-              testID="loginPasswordInput"
-              inputRef={passwordInputRef}
-              label={_(msg`Password`)}
-              autoCapitalize="none"
-              autoCorrect={false}
-              autoComplete="password"
-              returnKeyType="done"
-              enablesReturnKeyAutomatically={true}
-              secureTextEntry={true}
-              textContentType="password"
-              clearButtonMode="while-editing"
-              value={password}
-              onChangeText={setPassword}
-              onSubmitEditing={onPressNext}
-              blurOnSubmit={false} // HACK: https://github.com/facebook/react-native/issues/21911#issuecomment-558343069 Keyboard blur behavior is now handled in onSubmitEditing
-              editable={!isProcessing}
-              accessibilityHint={
-                identifier === ''
-                  ? _(msg`Input your password`)
-                  : _(msg`Input the password tied to ${identifier}`)
-              }
-            />
-            <TouchableOpacity
-              testID="forgotPasswordButton"
-              onPress={onPressForgotPassword}
-              accessibilityRole="button"
-              accessibilityLabel={_(msg`Forgot password`)}
-              accessibilityHint={_(msg`Opens password reset form`)}
-              style={[
-                a.rounded_sm,
-                t.atoms.bg_contrast_100,
-                {marginLeft: 'auto', left: 6, padding: 6},
-                a.z_10,
-              ]}>
-              <ButtonText style={t.atoms.text_contrast_medium}>
-                <Trans>Forgot?</Trans>
-              </ButtonText>
-            </TouchableOpacity>
-          </TextField.Root>
-        </View>
-        {error ? (
-          <View style={[styles.error, {marginHorizontal: 0}]}>
-            <Warning style={s.white} size="sm" />
-            <View style={(a.flex_1, a.ml_sm)}>
-              <Text style={[s.white, s.bold]}>{error}</Text>
-            </View>
-          </View>
-        ) : undefined}
-        <View style={[a.flex_row, a.align_center]}>
+        </TextField.Root>
+      </View>
+      <FormError error={error} />
+      <View style={[a.flex_row, a.align_center]}>
+        <Button
+          label={_(msg`Back`)}
+          variant="solid"
+          color="secondary"
+          size="small"
+          onPress={onPressBack}>
+          <ButtonText>
+            <Trans>Back</Trans>
+          </ButtonText>
+        </Button>
+        <View style={a.flex_1} />
+        {!serviceDescription && error ? (
           <Button
-            label={_(msg`Back`)}
+            testID="loginRetryButton"
+            label={_(msg`Retry`)}
+            accessibilityHint={_(msg`Retries login`)}
             variant="solid"
             color="secondary"
             size="small"
-            onPress={onPressBack}>
-            {_(msg`Back`)}
+            onPress={onPressRetryConnect}>
+            {_(msg`Retry`)}
           </Button>
-          <View style={s.flex1} />
-          {!serviceDescription && error ? (
-            <Button
-              testID="loginRetryButton"
-              label={_(msg`Retry`)}
-              accessibilityHint={_(msg`Retries login`)}
-              variant="solid"
-              color="secondary"
-              size="small"
-              onPress={onPressRetryConnect}>
-              {_(msg`Retry`)}
-            </Button>
-          ) : !serviceDescription ? (
-            <>
-              <ActivityIndicator />
-              <Text style={[t.atoms.text_contrast_high, a.pl_md]}>
-                <Trans>Connecting...</Trans>
-              </Text>
-            </>
-          ) : isProcessing ? (
+        ) : !serviceDescription ? (
+          <>
             <ActivityIndicator />
-          ) : isReady ? (
-            <Button
-              label={_(msg`Next`)}
-              accessibilityHint={_(msg`Navigates to the next screen`)}
-              variant="solid"
-              color="primary"
-              size="small"
-              onPress={onPressNext}>
-              {_(msg`Next`)}
-            </Button>
-          ) : undefined}
-        </View>
+            <Text style={[t.atoms.text_contrast_high, a.pl_md]}>
+              <Trans>Connecting...</Trans>
+            </Text>
+          </>
+        ) : isProcessing ? (
+          <ActivityIndicator />
+        ) : isReady ? (
+          <Button
+            label={_(msg`Next`)}
+            accessibilityHint={_(msg`Navigates to the next screen`)}
+            variant="solid"
+            color="primary"
+            size="small"
+            onPress={onPressNext}>
+            <ButtonText>
+              <Trans>Next</Trans>
+            </ButtonText>
+          </Button>
+        ) : undefined}
       </View>
-    </ScrollView>
+    </FormContainer>
   )
 }
